fix accural type

+++ internal/domain/balance/repository.go
@@ -4,5 +4,5 @@ import "context"
 
 type Repository interface {
 	GetByUserID(ctx context.Context, userID int) (*Balance, error)
-	Add(ctx context.Context, userID int, amount int64) error
+	Add(ctx context.Context, userID int, amount float64) error
 }
