+++ migrations/20250617160129_create_orders_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     number VARCHAR(255) NOT NULL UNIQUE,
     status VARCHAR(50) NOT NULL,
-    accrual INTEGER,
+    accrual DOUBLE PRECISION,
     uploaded_at TIMESTAMP WITH TIME ZONE NOT NULL,
     user_id INTEGER NOT NULL REFERENCES users(id)
 );
